+++ Sources/Scru128/Scru128Id.swift
@@ -7,6 +7,22 @@ public struct Scru128Id: LosslessStringConvertible {
   /// big-endian (network) byte order.
   public let bytes: [UInt8]
 
+  /// Creates an object from a byte tuple that represents a 128-bit unsigned integer.
+  ///
+  /// - Parameter bytes: A 16-byte byte array that represents a 128-bit unsigned integer in the
+  ///                    big-endian (network) byte order.
+  public init(
+    _ bytes: (
+      UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8,
+      UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8
+    )
+  ) {
+    self.bytes = [
+      bytes.0, bytes.1, bytes.2, bytes.3, bytes.4, bytes.5, bytes.6, bytes.7,
+      bytes.8, bytes.9, bytes.10, bytes.11, bytes.12, bytes.13, bytes.14, bytes.15,
+    ]
+  }
+
   /// Creates an object from a byte array that represents a 128-bit unsigned integer.
   ///
   /// - Parameter byteArray: A 16-byte byte array that represents a 128-bit unsigned integer in the