+++ Tests/Scru128Tests/Scru128IdTests.swift
@@ -96,6 +96,15 @@ final class Scru128IdTests: XCTestCase {
       XCTAssertEqual(Scru128Id(e.byteArray), e)
       XCTAssertEqual(Scru128Id(e.timestamp, e.counterHi, e.counterLo, e.entropy), e)
       XCTAssertEqual(try decoder.decode(Scru128Id.self, from: try encoder.encode(e)), e)
+
+      let ar = e.byteArray
+      XCTAssertEqual(
+        Scru128Id(
+          (
+            ar[0], ar[1], ar[2], ar[3], ar[4], ar[5], ar[6], ar[7],
+            ar[8], ar[9], ar[10], ar[11], ar[12], ar[13], ar[14], ar[15]
+          )
+        ), e)
     }
   }